chore(engine): migrate ExpressionIncidentTest
